fix(tool): separate tool from config
